requirements.txt
transformers==4.38.*
beir==1.0.1
datasets==2.14.1
tqdm==4.67.*
scipy==1.13.*
evaluate==0.2.2
spacy==3.7.2
accelerate==1.10.*
elasticsearch==7.17.9
sentence-transformers==5.1.*
faiss-cpu==1.9.*